Rename hasNumActivities to minNumActivities
