+++ __tests__/setup/conditions/minNumActivities.js
